chore: move translations to preferences folder
