+++ features/HomeTab/AdvancedFilter/RoomsAndBeds/RoomsAndBeds.tsx
@@ -2,7 +2,7 @@ import { useContext } from "react";
 
 import RoomsAndBedsItem from "./RoomsAndBedsItem/RoomsAndBedsItem";
 import SettingsContext from "../../../../contexts/SettingsContext";
-import translations from "../../../../translations/translations";
+import translations from "../../../../preferences/translations";
 import { ButtonsType } from "../AdvancedFilter";
 import FilterItem from "../FilterItem/FilterItem";
