+++ features/ProfileTab/MainScreen/MainScreen.tsx
@@ -8,7 +8,7 @@ import SectionItem from "./SectionItem/SectionItem";
 import SafeAreaScreenWrapper from "../../../components/SafeAreaScreenWrapper/SafeAreaScreenWrapper";
 import AuthContext from "../../../contexts/AuthContext";
 import SettingsContext from "../../../contexts/SettingsContext";
-import translations from "../../../translations/translations";
+import translations from "../../../preferences/translations";
 
 interface MainScreenProps {
   route: any;
