+++ features/ProfileTab/ProfileScreen/EditableEmail/EditableEmail.tsx
@@ -3,7 +3,7 @@ import { View } from "react-native";
 import { Button, Surface, Text, TextInput } from "react-native-paper";
 
 import SettingsContext from "../../../../contexts/SettingsContext";
-import translations from "../../../../translations/translations";
+import translations from "../../../../preferences/translations";
 
 const EditableEmail = () => {
   const { settings, setSettings } = useContext(SettingsContext);
